tweaks to various modules:
* mod_alias: add %u format for url-encoded version of arg
* mod_markov: allow the markov_gen mod_msg to specify a length
* mod_haiku: multi-line support with INSOBOT_MULTILINE_HAIKU env var
* mod_twitch: better vod/alias and tag_list logic
* mod_automod: add cast to time_t printf to avoid warning
* mod_info, mod_bf: add global flag

+++ src/mod_automod.c
@@ -274,7 +274,7 @@ static int am_score_links(const Suspect* s, const char* msg, size_t len){
 			time_t user_created_date = 0;
 			MOD_MSG(ctx, "twitch_get_user_date", s->name, &get_user_cb, &user_created_date);
 
-			printf("twitch user time: %zd\n", (now - user_created_date));
+			printf("twitch user time: %zd\n", (size_t)(now - user_created_date));
 
 			if((now - user_created_date) < (24*60*60)){
 				return 500;
